In this commit the entire API logic was built. The logic include POST, GET, DELETE request to Postgres DB.
.env - Store the hardcoded values of main properties such as DB url, DB username, main table name and so on.

.gitignore - ignore file for unwanted files when doing commits.

requirement.txt - list of packages that needed to be installed in order for the app to work.

backservices.py - the main part of our api app. In here we will deal with the API different kinds of request.

helper.py - a helping class which usually deal with the data of the request and not the request itself.

home.html - The main page of the app. What the user will see without doing an API request